# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BuildOS API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  builder: "builder",
  client: "client",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  createdAt: string;
}

export type RegisterBodyRole =
  (typeof RegisterBodyRole)[keyof typeof RegisterBodyRole];

export const RegisterBodyRole = {
  builder: "builder",
  client: "client",
} as const;

export interface RegisterBody {
  name: string;
  email: string;
  password: string;
  role: RegisterBodyRole;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  planning: "planning",
  active: "active",
  on_hold: "on_hold",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Project {
  id: number;
  builderId: number;
  /** @nullable */
  clientId: number | null;
  clientName: string;
  /** @nullable */
  clientEmail: string | null;
  name: string;
  address: string;
  status: ProjectStatus;
  /** @nullable */
  startDate: string | null;
  /** @nullable */
  notes: string | null;
  progress: number;
  createdAt: string;
  updatedAt: string;
}

export type CreateProjectBodyStatus =
  (typeof CreateProjectBodyStatus)[keyof typeof CreateProjectBodyStatus];

export const CreateProjectBodyStatus = {
  planning: "planning",
  active: "active",
  on_hold: "on_hold",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface CreateProjectBody {
  clientName: string;
  /** @nullable */
  clientEmail?: string | null;
  name: string;
  address: string;
  status: CreateProjectBodyStatus;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  notes?: string | null;
  progress?: number;
}

/**
 * @nullable
 */
export type UpdateProjectBodyStatus =
  | (typeof UpdateProjectBodyStatus)[keyof typeof UpdateProjectBodyStatus]
  | null;

export const UpdateProjectBodyStatus = {
  planning: "planning",
  active: "active",
  on_hold: "on_hold",
  completed: "completed",
  cancelled: "cancelled",
  null: "null",
} as const;

export interface UpdateProjectBody {
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  clientEmail?: string | null;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  status?: UpdateProjectBodyStatus;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  progress?: number | null;
}

export type ContractStatus =
  (typeof ContractStatus)[keyof typeof ContractStatus];

export const ContractStatus = {
  draft: "draft",
  sent: "sent",
  signed: "signed",
} as const;

export interface Contract {
  id: number;
  projectId: number;
  title: string;
  /** @nullable */
  fileUrl: string | null;
  /** @nullable */
  version: string | null;
  status: ContractStatus;
  uploadedAt: string;
}

export type CreateContractBodyStatus =
  (typeof CreateContractBodyStatus)[keyof typeof CreateContractBodyStatus];

export const CreateContractBodyStatus = {
  draft: "draft",
  sent: "sent",
  signed: "signed",
} as const;

export interface CreateContractBody {
  title: string;
  /** @nullable */
  fileUrl?: string | null;
  /** @nullable */
  version?: string | null;
  status: CreateContractBodyStatus;
}

/**
 * @nullable
 */
export type UpdateContractBodyStatus =
  | (typeof UpdateContractBodyStatus)[keyof typeof UpdateContractBodyStatus]
  | null;

export const UpdateContractBodyStatus = {
  draft: "draft",
  sent: "sent",
  signed: "signed",
  null: "null",
} as const;

export interface UpdateContractBody {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  fileUrl?: string | null;
  /** @nullable */
  version?: string | null;
  /** @nullable */
  status?: UpdateContractBodyStatus;
}

export type ChangeOrderStatus =
  (typeof ChangeOrderStatus)[keyof typeof ChangeOrderStatus];

export const ChangeOrderStatus = {
  draft: "draft",
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface ChangeOrder {
  id: number;
  projectId: number;
  title: string;
  /** @nullable */
  description: string | null;
  amount: number;
  status: ChangeOrderStatus;
  createdBy: number;
  /** @nullable */
  approvedBy: number | null;
  createdAt: string;
  /** @nullable */
  approvedAt: string | null;
}

export type CreateChangeOrderBodyStatus =
  (typeof CreateChangeOrderBodyStatus)[keyof typeof CreateChangeOrderBodyStatus];

export const CreateChangeOrderBodyStatus = {
  draft: "draft",
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface CreateChangeOrderBody {
  title: string;
  /** @nullable */
  description?: string | null;
  amount: number;
  status: CreateChangeOrderBodyStatus;
}

/**
 * @nullable
 */
export type UpdateChangeOrderBodyStatus =
  | (typeof UpdateChangeOrderBodyStatus)[keyof typeof UpdateChangeOrderBodyStatus]
  | null;

export const UpdateChangeOrderBodyStatus = {
  draft: "draft",
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
  null: "null",
} as const;

export interface UpdateChangeOrderBody {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  amount?: number | null;
  /** @nullable */
  status?: UpdateChangeOrderBodyStatus;
}

export interface Photo {
  id: number;
  projectId: number;
  fileUrl: string;
  /** @nullable */
  caption: string | null;
  visibleToClient: boolean;
  uploadedBy: number;
  createdAt: string;
}

export interface CreatePhotoBody {
  fileUrl: string;
  /** @nullable */
  caption?: string | null;
  visibleToClient?: boolean;
}

export interface UpdatePhotoBody {
  /** @nullable */
  caption?: string | null;
  /** @nullable */
  visibleToClient?: boolean | null;
}

export interface ActivityLog {
  id: number;
  projectId: number;
  type: string;
  description: string;
  /** @nullable */
  createdBy: number | null;
  createdAt: string;
}