# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BuildOS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /projects/{projectId}/change-orders:
    get:
      operationId: listChangeOrders
      tags: [changeOrders]
      summary: List change orders for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of change orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChangeOrder"
    post:
      operationId: createChangeOrder
      tags: [changeOrders]
      summary: Create a change order
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChangeOrderBody"
      responses:
        "201":
          description: Change order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChangeOrder"
  /projects/{projectId}/change-orders/{id}:
    patch:
      operationId: updateChangeOrder
      tags: [changeOrders]
      summary: Update a change order
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateChangeOrderBody"
      responses:
        "200":
          description: Change order updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChangeOrder"
    delete:
      operationId: deleteChangeOrder
      tags: [changeOrders]
      summary: Delete a change order
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{projectId}/change-orders/{id}/approve:
    post:
      operationId: approveChangeOrder
      tags: [changeOrders]
      summary: Approve a change order
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Change order approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChangeOrder"
  /projects/{projectId}/change-orders/{id}/reject:
    post:
      operationId: rejectChangeOrder
      tags: [changeOrders]
      summary: Reject a change order
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Change order rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChangeOrder"
...
  /projects/{projectId}/activity:
    get:
      operationId: listActivity
      tags: [activity]
      summary: List activity logs for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Activity log
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityLog"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [builder, client]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - createdAt
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [builder, client]
      required:
        - name
        - email
        - password
        - role
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
...
    ChangeOrder:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        amount:
          type: number
        status:
          type: string
          enum: [draft, pending, approved, rejected]
        createdBy:
          type: integer
        approvedBy:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        approvedAt:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - projectId
        - title
        - description
        - amount
        - status
        - createdBy
        - approvedBy
        - createdAt
        - approvedAt
    CreateChangeOrderBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        amount:
          type: number
        status:
          type: string
          enum: [draft, pending, approved, rejected]
      required:
        - title
        - amount
        - status
    UpdateChangeOrderBody:
      type: object
      properties:
        title:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        amount:
          type: ["number", "null"]
        status:
          type: ["string", "null"]
          enum: [draft, pending, approved, rejected, "null"]
...
    ActivityLog:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        type:
          type: string
        description:
          type: string
        createdBy:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - type
        - description
        - createdBy
        - createdAt