# lib/api-zod/src/generated/types/activityLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BuildOS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityLog {
  id: number;
  projectId: number;
  type: string;
  description: string;
  /** @nullable */
  createdBy: number | null;
  createdAt: Date;
}