# lib/api-zod/src/generated/types/changeOrder.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BuildOS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChangeOrderStatus } from "./changeOrderStatus";

export interface ChangeOrder {
  id: number;
  projectId: number;
  title: string;
  /** @nullable */
  description: string | null;
  amount: number;
  status: ChangeOrderStatus;
  createdBy: number;
  /** @nullable */
  approvedBy: number | null;
  createdAt: Date;
  /** @nullable */
  approvedAt: Date | null;
}