# lib/api-zod/src/generated/types/createChangeOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BuildOS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateChangeOrderBodyStatus } from "./createChangeOrderBodyStatus";

export interface CreateChangeOrderBody {
  title: string;
  /** @nullable */
  description?: string | null;
  amount: number;
  status: CreateChangeOrderBodyStatus;
}